+++ tests/test_xls.py
@@ -1,7 +1,9 @@
+from io import BytesIO
 from pathlib import Path
 
 import anyio
 import pytest
+from fastapi import UploadFile
 
 from asynctor.xls import df_to_datas, load_xls, read_excel
 
@@ -9,10 +11,14 @@ from asynctor.xls import df_to_datas, load_xls, read_excel
 @pytest.mark.anyio
 async def test_read():
     demo = Path(__file__).parent / "demo.xlsx"
+    content = demo.read_bytes()
     df = await read_excel(demo)
-    df2 = await read_excel(demo.read_bytes())
+    df2 = await read_excel(content)
     df3 = await read_excel(anyio.Path(demo))
+    df4 = await read_excel(BytesIO(content))
+    df5 = await read_excel(UploadFile(BytesIO(content)))
     assert df2.compare(df).empty and df3.compare(df).empty
+    assert df4.compare(df).empty and df5.compare(df).empty
     data = df_to_datas(df)
     assert data == [
         {"Column1": "row1-\\t%c", "Column2\nMultiLines": 0, "Column 3": 1, 4: ""},